+++ tools4a-context/src/androidTest/java/com/github/panpf/tools4a/context/test/ContextxTest.java
@@ -47,7 +47,7 @@ public class ContextxTest {
 
     @Rule
     @NonNull
-    public final ActivityTestRule getActivityRule() {
+    public final ActivityTestRule<TestActivity> getActivityRule() {
         return this.activityRule;
     }